Fix init calls

+++ src/taurex/data/profiles/chemistry/gas/twolayergas.py
@@ -273,7 +273,7 @@ class TwoLayerGas(Gas):
 
         c_smooth = 10 ** movingaverage(np.log10(chemprofile), int(wsize))
 
-        border = np.int64((len(chemprofile) - len(c_smooth)) / 2)
+        border = int((len(chemprofile) - len(c_smooth)) / 2)
 
         self._mix_profile = chemprofile[::-1]
